Run dependency rules as dependencies are added to the graph
Previously, dependency rules (resolve/substitution) were applied in two
places: once when building the 'pending nodes map', and again when
resolving the dependency (DependencySubstitutionResolver).
These 2 processes have been replaced by a single process that runs any
dependency rules at the time the dependency is added to the graph.
This simplifies the code and ensures that what ends up in the graph
is always the result of executing the dependency rules.

Moved dependency substitution into one place