Add a specialized version of registerConflict when there's only one target
In case of numerous conflicts, for example if lots of pending dependencies are added and
they eventually become used, then conflict resolution handling becomes more visible in
performance traces, because of the aggressive creation of sets when they are not necessary.

This commit removes the need to create sets to handle the different cases.